Move filters to sub packages, add comment for CorsFilter class
