+++ src/main/java/edu/harvard/iq/dataverse/api/util/JsonResponseBuilder.java
@@ -1,7 +1,5 @@
 package edu.harvard.iq.dataverse.api.util;
 
-import edu.harvard.iq.dataverse.api.ApiBlockingFilter;
-
 import jakarta.json.Json;
 import jakarta.json.JsonValue;
 import jakarta.json.JsonObjectBuilder;
@@ -12,6 +10,8 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 
 import org.apache.commons.lang3.exception.ExceptionUtils;
+
+import edu.harvard.iq.dataverse.api.filter.ApiBlockingFilter;
 
 import java.io.IOException;
 import java.util.Optional;
